rename: BTS-211 / 스터디노트 타입 파일명 변경
