+++ src/features/dashboard/studynote/detail/service/query.ts
@@ -1,4 +1,4 @@
-import type { StudyNoteDetail } from '@/features/dashboard/studynote/detail/tyoe';
+import type { StudyNoteDetail } from '@/features/dashboard/studynote/detail/type';
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
 
 import { getStudyNoteDetailOption } from './query-options';
