+++ src/features/dashboard/studynote/detail/type.ts
